Attempted fix for constant database operations
Attempted fix to remove the constant database operations triggered whenever the staff is updated. The way it should work is that it caches the staff description into the spellcasting entry flags, then whenever the staff is updated, it compares the cached and current descriptions. If they match, nothing the module cares about should have changed, and assume there's no need to update the spellcasting entry. This should prevent the constant database operations whenever eg. the staff changes grip.